+++ src/main/kotlin/com/cornerstone/cheque/model/Transaction.kt
@@ -1,6 +1,7 @@
 package com.cornerstone.cheque.model
 
 import jakarta.persistence.*
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Entity
@@ -17,13 +18,13 @@ data class Transaction(
     @JoinColumn(name = "receiver_account_id", referencedColumnName = "account_number")
     val receiverAccount: Account,
 
-    val amount: Double,
+    val amount: BigDecimal,
     val createdAt: LocalDate
 )
 
 data class TransactionRequest(
     val senderAccount: String,
     val receiverAccount: String,
-    val amount: Double
+    val amount: BigDecimal
 )
